it can load localhost
