+++ local.log
@@ -1,0 +1,14 @@
+
+
+BrowserStackLocal v6.9
+
+Setting output as verbose
+You can now access your local server(s) in our remote browser.
+
+
+Connecting to BrowserStack using WebSocket protocol...
+Platform supports WebSocket connections.
+
+Press Ctrl-C to exit
+
+
