+++ test/getCapabilityText/index.js
@@ -1,0 +1,14 @@
+const isString = require('j1/isString');
+
+function getCapabilityText(capability) {
+	const fields = [];
+	for (const key of ['os', 'os_version', 'browser', 'browser_version', 'resolution']) {
+		const value = capability[key];
+		if (isString(value)) {
+			fields.push(value);
+		}
+	}
+	return fields.join(',');
+}
+
+module.exports = getCapabilityText;
