+++ test/index.js
@@ -3,20 +3,10 @@ const SableServer = require('sable');
 const browserstack = require('browserstack-local');
 const {Builder} = require('selenium-webdriver');
 const $console = require('j1/console').create('test');
-const isString = require('j1/isString');
 const promisify = require('j1/promisify');
 const useBrowserStack = require('./useBrowserStack');
-
-function getCapabilityText(capability) {
-	const fields = [];
-	for (const key of ['os', 'os_version', 'browser', 'browser_version', 'resolution']) {
-		const value = capability[key];
-		if (isString(value)) {
-			fields.push(value);
-		}
-	}
-	return fields.join(',');
-}
+const getCapabilityText = require('./getCapabilityText');
+const waitRun = require('./waitRun');
 
 async function test(capabilities = require('./capabilities')) {
 	const capability = capabilities.shift();
@@ -24,7 +14,17 @@ async function test(capabilities = require('./capabilities')) {
 		return;
 	}
 	const console = $console.create(getCapabilityText(capability));
-	capability.localIdentifier = exports.localIdentifier;
+	if (useBrowserStack) {
+		Object.assign(
+			capability,
+			{
+				'browserstack.local': true,
+				'browserstack.localIdentifier': exports.localIdentifier,
+				'browserstack.user': useBrowserStack.user,
+				'browserstack.key': useBrowserStack.key
+			}
+		)
+	}
 	const builder = new Builder()
 	.withCapabilities(capability);
 	if (useBrowserStack) {
@@ -45,7 +45,7 @@ async function onEnd() {
 		await exports.server.close();
 	}
 	if (exports.bsLocal) {
-		await exports.bsLocal.stop();
+		await promisify(exports.bsLocal.stop, exports.bsLocal)();
 	}
 	if (exports.driver) {
 		await exports.driver.quit();
@@ -71,13 +71,14 @@ Promise.resolve()
 		$console.info(`start a browserstack-local ${port}`);
 		await promisify(bsLocal.start, bsLocal)({
 			// https://github.com/browserstack/browserstack-local-nodejs/blob/master/lib/Local.js
+			key: useBrowserStack.key,
 			verbose: true,
 			forceLocal: true,
 			onlyAutomate: true,
-			onlyCommand: true,
 			only: `localhost,${port},0`,
 			localIdentifier: exports.localIdentifier
 		});
+		await waitRun(bsLocal);
 	}
 	$console.info('start test');
 	await test();
